Update Final Fantasy XIV package derivation
Download the app from the Sparkle update site instead of the main site,
so we’re on the latest version without having the installation be
replaced as soon as it is launched. This avoids a situation where
`darwin-rebuild switch` would rebuild `~/Applications/Home Manage Apps`,
and revert the updated version. Also, set the version to the same one
used by Sparkle and in the app itself. This is independent of the game’s
version as well as the launcher version.