Fix Unbound service start failures
Unbound needs to be built with systemd support, which it is not by
default in NixOS.  The Unbound service actually uses its own variant
that is, but the overlay replaced it with one that was not.  Change the
overlay to update `unbound-with-systemd` to include DoH support instead
of using a broken variant.