Use flake-utils-plus for the flake
This is a massive refactor.  The hand-rolled flake has been replaced
with [flake-utils-plus][1], which provides some extra configuration
options (such as being able to manage NIX_PATH with the flake).  I’m not
currently taking advantage of it, but that could change in the future.
It also makes it a bit easier to define and manage overlays since that’s
all part of the provided builders.

For the most part, the hierarchy hasn’t changed.  Overlays don’t seem to
work unless you define them via the `overlayBuilder`. I also reorganized
and cleaned up the `common` folder.  Individual hosts should no longer
reference files in `common`.  Those files are included in the modules
automatically.  Additionally, it is now possible to define
platform-specific common home-manager configuration that spans across
architectures (in anticipation of getting an Apple Silicon Mac in the
future).

[1]: https://github.com/gytis-ivaskevicius/flake-utils-plus/